Add concurrent execution of scheduled commands, split into main and storage files
For this, each command gets a random UUID version 4 on adding it and we track
its state on disk to not start commands that are currently running.
Access to the command store file is controlled with a lock in the main daemon
and its goroutines, but when adding commands with another process, there
could still be conflicts that need to be solved with synchronizin the file access
across processes.