"Add recurring commands function, read from config files (toml)
Currently, all toml files in current directory are read as commands
to regularly execute with a certain time elapsing between runs.
The file name serves as a unique identifier for
the command with its arguments and the duration between consecutive executions.
The add command CLI is not yet adapted, as the focus is on control through config
files first.

If a file name is not present anymore, the state (last run and success of that run)
will be deleted.
If only config file *contents* change, its values are updated accordingly in the program
state and the last run und success of that run are kept.

Old identification of commands via UUID v4 not yet removed, but not needed anymore.

Extracted a few helper functions in own file.